Patitas Felices: Frontend: Clening of dogs actions/reducer; Creation of request actions and reducers;
