+++ frontend/src/_redux/reducer/dogsReducer.js
@@ -16,7 +16,6 @@ const initialState = {
 	setTemperaments: 'All',
 	currentPage: 1,
 	maybeAdoptedDog: {},
-	allAdoptions: {},
 };
 
 export const dogsSlice = createSlice({
@@ -135,11 +134,6 @@ export const dogsSlice = createSlice({
 			state.maybeAdoptedDog = dog;
 		},
 
-
-		setAdoptionDog: (state, action) => {
-			state.allAdoptions = action.payload
-		},
-
 		emptyMaybeAdoptedDog: (state) => {
 			state.maybeAdoptedDog = {};
 		},
@@ -161,7 +155,6 @@ export const {
 	setEmptyDetail,
 	setPages,
 	setMaybeAdoptedDog,
-	setAdoptionDog,
 	emptyMaybeAdoptedDog,
 } = dogsSlice.actions;
