+++ frontend/src/_redux/reducer/requestReducer.js
@@ -1,0 +1,25 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialState = {
+	allAdoptions: [],
+};
+
+
+export const requestSlice = createSlice({
+	name: 'request',
+	initialState,
+	reducers: {
+
+        setAdoptionDog: (state, action) => {
+            state.allAdoptions = action.payload
+        },
+    }
+});
+
+export const {
+
+            setAdoptionDog,
+
+        } = requestSlice.actions;
+
+export default requestSlice.reducer;
